pyproject.toml
[project]
name = "harbor-clerk"
version = "0.4.1"
description = "Harbor Clerk - document dropbox with hybrid retrieval"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "pgvector>=0.3.6",
    "minio>=7.2.0",
    "bcrypt>=4.2.0",
    "httpx>=0.28.0",
    "PyJWT>=2.9.0",
    "python-multipart>=0.0.18",
    "pytesseract>=0.3.13",
    "pypdfium2>=4.0.0",
    "langdetect>=1.0.9",
    "psycopg2-binary>=2.9.0",
    "Pillow>=10.0.0",
    "mcp>=1.9.0",
    "huggingface_hub>=0.25.0",
    "spacy>=3.7.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/harbor_clerk"]

[project.scripts]
harbor-clerk-api = "harbor_clerk.api.app:main"
harbor-clerk-worker = "harbor_clerk.worker.entry:main"
harbor-clerk-seed = "harbor_clerk.seed:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "E501",   # line length handled by formatter
    "B008",   # Depends() in args is standard FastAPI
    "B904",   # raise from None not needed for HTTPException
    "B905",   # zip strict= not needed everywhere
    "UP042",  # StrEnum migration risks SQLAlchemy enum compat
    "SIM108", # ternary not always more readable
    "SIM105", # contextlib.suppress is stylistic preference
]

[tool.ruff.format]
quote-style = "double"
